requirements.txt
Pillow==10.0.0
tk==0.1.0